Cleanup ImperativeHandle logic
